requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.8
django-autoslug-iplweb==1.9.4.dev0
django-cities==0.6
<<<<<<< HEAD
django-model-utils==4.0.0
djangorestframework==3.11.0
GDAL @ file:///C:/Users/Damis/PycharmProjects/location_api/gdal/GDAL-3.1.2-cp38-cp38-win_amd64.whl
=======
django-cors-headers==3.4.0
django-heroku==0.3.1
django-model-utils==4.0.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
inflection==0.5.0
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
progress==1.5
>>>>>>> 775b3267cd8f6e6808191192f1cc1176a02b0306
progressbar2==3.51.4
pyparsing==2.4.6
python-dotenv==0.14.0
python-utils==2.4.0
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
tqdm==4.48.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
djangorestframework-jwt~=1.11.0
